Final_Agentic-AI-Based-HR-Compliance-Assistant/requirements.txt
faiss-cpu
numpy
pandas
scikit-learn
joblib

sentence-transformers
langchain
langchain-community
langchain-core
langgraph

duckdb
rapidfuzz
sqlglot
requests

pypdf
docx2txt

fastapi 
uvicorn 
streamlit 
requests